Move `typeck` logically in the module tree out to the root and clamp down on its exports. Remove some dead code that is revealed.
