+++ src/librustc/middle/typeck/check/closure.rs
@@ -17,8 +17,8 @@ use super::FnCtxt;
 use middle::infer;
 use middle::subst;
 use middle::ty::{mod, Ty};
-use middle::typeck::astconv;
-use middle::typeck::rscope::RegionScope;
+use typeck::astconv;
+use typeck::rscope::RegionScope;
 use syntax::abi;
 use syntax::ast;
 use syntax::ast_util;
