+++ src/librustc/middle/typeck/check/method/mod.rs
@@ -15,11 +15,11 @@ use middle::subst::{Subst};
 use middle::traits;
 use middle::ty::*;
 use middle::ty;
-use middle::typeck::astconv::AstConv;
-use middle::typeck::check::{FnCtxt};
-use middle::typeck::check::{impl_self_ty};
-use middle::typeck::check::vtable;
-use middle::typeck::check::vtable::select_new_fcx_obligations;
+use typeck::astconv::AstConv;
+use typeck::check::{FnCtxt};
+use typeck::check::{impl_self_ty};
+use typeck::check::vtable;
+use typeck::check::vtable::select_new_fcx_obligations;
 use middle::infer;
 use util::ppaux::{Repr, UserString};
