+++ src/librustc/middle/typeck/check/method/probe.rs
@@ -19,8 +19,8 @@ use middle::traits;
 use middle::ty::{mod, Ty};
 use middle::ty::{MethodObject};
 use middle::ty_fold::HigherRankedFoldable;
-use middle::typeck::check;
-use middle::typeck::check::{FnCtxt, NoPreference};
+use typeck::check;
+use typeck::check::{FnCtxt, NoPreference};
 use middle::infer;
 use middle::infer::InferCtxt;
 use syntax::ast;
