+++ src/librustc/middle/typeck/check/regionck.rs
@@ -120,10 +120,10 @@ use middle::region::CodeExtent;
 use middle::traits;
 use middle::ty::{ReScope};
 use middle::ty::{mod, Ty, MethodCall};
-use middle::typeck::astconv::AstConv;
-use middle::typeck::check::FnCtxt;
-use middle::typeck::check::regionmanip;
-use middle::typeck::check::vtable;
+use typeck::astconv::AstConv;
+use typeck::check::FnCtxt;
+use typeck::check::regionmanip;
+use typeck::check::vtable;
 use middle::infer::resolve_and_force_all_but_regions;
 use middle::infer::resolve_type;
 use middle::infer;
