+++ src/librustc/middle/typeck/collect.rs
@@ -43,13 +43,13 @@ use middle::ty::{ImplContainer, ImplOrTraitItemContainer, TraitContainer};
 use middle::ty::{Polytype};
 use middle::ty::{mod, Ty};
 use middle::ty_fold::TypeFolder;
-use middle::typeck::astconv::{AstConv, ty_of_arg};
-use middle::typeck::astconv::{ast_ty_to_ty, ast_region_to_region};
-use middle::typeck::astconv;
+use typeck::astconv::{AstConv, ty_of_arg};
+use typeck::astconv::{ast_ty_to_ty, ast_region_to_region};
+use typeck::astconv;
 use middle::infer;
-use middle::typeck::rscope::*;
-use middle::typeck::{CrateCtxt, lookup_def_tcx, no_params, write_ty_to_tcx};
-use middle::typeck;
+use typeck::rscope::*;
+use typeck::{CrateCtxt, lookup_def_tcx, no_params, write_ty_to_tcx};
+use typeck;
 use util::nodemap::{FnvHashMap, FnvHashSet};
 use util::ppaux;
 use util::ppaux::{Repr,UserString};
